added some improvements 2

+++ src/main/java/com/martyx/employeemanager/service/EmployeeServiceBean.java
@@ -1,6 +1,6 @@
 package com.martyx.employeemanager.service;
 
-import com.martyx.employeemanager.exception.UserNotFoundException;
+import com.martyx.employeemanager.exception.EmployeeNotFoundException;
 import com.martyx.employeemanager.model.Employee;
 import com.martyx.employeemanager.repo.EmployeeRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -27,7 +27,7 @@ public class EmployeeServiceBean implements EmployeeService {
 
     @Override
     public Employee findEmployeeById(Integer id) {
-        return employeeRepository.findById(id).orElseThrow(() -> new UserNotFoundException("User was not found"));
+        return employeeRepository.findById(id).orElseThrow(() -> new EmployeeNotFoundException("Employee was not found"));
     }
 
     @Override
@@ -42,6 +42,10 @@ public class EmployeeServiceBean implements EmployeeService {
 
     @Override
     public void deleteEmployee(Integer id) {
-        employeeRepository.deleteById(id);
+        if (id == null) {
+            throw new EmployeeNotFoundException("Employee with id not exist");
+        } else {
+            employeeRepository.deleteById(id);
+        }
     }
 }
